% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{agraph}
\alias{agraph}
\title{Segmentation using graph structure}
\usage{
agraph(
  gamma,
  graph,
  lambda = 10^seq(-4, 4, length.out = 50),
  weights = NULL,
  shrinkage = TRUE,
  delta = 1e-10,
  tol = 1e-08,
  thresh = 0.01,
  itermax = 50000
)
}
\arguments{
\item{gamma}{entry vector to regularize}

\item{graph}{graph (an \url{igraph} object) giving the regularization structure}

\item{lambda}{regularizing constant}

\item{weights}{weights for gamma. Default value is one.}

\item{delta}{Computational constant in the adaptive ridge reweighting formula.}

\item{tol}{Tolerance to test for convergence of the adaptive ridge}

\item{thresh}{Thresholding constant used to fuse two adjacent regions with close value of \code{gamma}.}

\item{itermax}{Total number of iterations. Default value is 10000. Setting a low value can make the procedure return NULL entries for some values of \code{lambda}.}
}
\description{
Segmentation using graph structure
}
