% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{agraph_one_lambda}
\alias{agraph_one_lambda}
\title{Segmentation using graph structure}
\usage{
agraph_one_lambda(
  gamma,
  graph,
  lambda = 1,
  weights = NULL,
  delta = 1e-10,
  tol = 1e-08,
  thresh = 0.01
)
}
\arguments{
\item{gamma}{entry vector to regularize}

\item{graph}{an igraph object (from package \code{igraph}) giving the regularization structure}

\item{lambda}{regularizing constant}

\item{weights}{weights for gamma. Default value is one.}

\item{delta}{Computational constant in the adaptive ridge reweighting formula.}

\item{tol}{Tolerance to test for convergence of the adaptive ridge}

\item{thresh}{Thresholding constant used to fuse two adjacent regions with close value of \code{gamma}.}
}
\description{
Segmentation using graph structure
}
