% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{build_jt}
\alias{build_jt}
\title{Build junction tree of a graph}
\usage{
build_jt(g, verbose = FALSE)
}
\arguments{
\item{g}{graph}

\item{verbose}{whether to print progress while running. Defaults to \code{FALSE}.}
}
\description{
Build junction tree of a graph
}
