% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departement}
\alias{departement}
\title{French departement}
\format{
An \code{\link[sf]{sf}} object, with lattitude and longitude
}
\source{
The data set comes from \url{https://github.com/gregoiredavid/france-geojson/}, under the "Licence Ouverte / Open Licence v2.0" licence.
}
\usage{
departement
}
\description{
Geographical data of the French administrative units "département"
This data set excludes overseas departement as well as the two departements of the island of Corsica.
}
\examples{
data(departement)
plot(departement["geometry"])
}
\keyword{datasets}
