% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{flsa_graph}
\alias{flsa_graph}
\title{Segmentation using graph structure and the fused lasso estimate}
\usage{
flsa_graph(gamma, graph, lambda)
}
\arguments{
\item{gamma}{entry vector to regularize}

\item{graph}{graph (an \link[igraph]{igraph} object) giving the regularization structure}

\item{lambda}{regularizing constant}
}
\description{
Wrapper around the function \code{flsa::flsa}.
}
\references{
Hoefling, H., A Path Algorithm for the Fused Lasso Signal Approximator,
Journal of Computational and Graphical Statistics (2010)
\doi{10.1198/jcgs.2010.09208}
}
