% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{graph_aridge}
\alias{graph_aridge}
\title{Regularization of spatial data using adjacency graph}
\usage{
graph_aridge(gamma, sigma_sq, adj, pen = 10^seq(-4, 4, length = 100),
  maxiter = 10000, epsilon = 1e-05, thresh = 1e-08)
}
\arguments{
\item{gamma}{vector of estimated values}

\item{sigma_sq}{vector of estimated variances}

\item{adj}{adjacency matrix of the regions corresponding to each values}

\item{pen}{penalty value parameter}

\item{maxiter}{maximal number of iterations in the adaptive ridge algorithm, see details.}

\item{epsilon}{numerical constant used in the adaptive ridge algorithm. Should be small compared to \code{gamma} and large compared to machine precision.}

\item{thresh}{relative tolerance for the convergence of the adaptive ridge iterations.}
}
\description{
Regularization of spatial data using adjacency graph
}
