% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphseg-package.R
\docType{package}
\name{graphseg-package}
\alias{graphseg}
\alias{graphseg-package}
\title{graphseg: A package for piece-wise constant estimation of graph-based signals}
\description{
The main function of this package is \code{agraph}, which performs segmentation
of signals on graphs into a piecewise-constant signal.
(the function \code{graph_aridge} is deprecated and kept here for backward
compatibility).
}
\details{
See the corresponding paper: \doi{10.48550/arXiv.2206.06752}
}
\seealso{
The \code{\link{flsa}} package on CRAN, which performs a similar task.
}
\author{
\strong{Maintainer}: Vivien Goepp \email{vivien.goepp@gmail.com} (\href{https://orcid.org/0000-0001-6961-4260}{ORCID})

}
\keyword{internal}
