% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphseg-package.R
\docType{package}
\name{graphseg}
\alias{graphseg}
\title{graphseg: A package for piece-wise constant estimation of graph-based signal}
\description{
The main functino of this package is the fuction \code{agraph}
(the function \code{graph_aridge} is deprecated and kept here for backward
compatibility).
}
\seealso{
The \code{flsa} package on CRAN, which performs a similar task
}
