% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{loglik}
\alias{loglik}
\alias{hessian}
\alias{score}
\title{this is a title}
\usage{
loglik(par, gamma, sigma_sq, K, pen)

hessian(par, gamma, sigma_sq, K, pen)

score(par, gamma, sigma_sq, K, pen)
}
\arguments{
\item{par}{current parameter.}

\item{gamma}{observed signal on a graph.}

\item{sigma_sq}{variance of \code{gamma} (functions as weights for \code{gamma}).}

\item{K}{weighted laplacian matrix of the graph.}

\item{pen}{penalty constant (must be a positive number). Higher values of \code{pen} yields a more regularized fit.}
}
\value{
The penalized negative log-likelihood of the parameter (function \code{loglik}), or its first (function \code{score}) and second (function \code{hessian}) derivatives.
}
\description{
Penalized log-likelihood, score, and hessian of the graph-based signal.
}
\author{
Vivien Goepp \email{vivien.goepp@gmail.com}
}
