% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris}
\alias{paris}
\title{IRIS of Paris}
\format{
An \link[sf]{sf} object, with lattitude and longitude
}
\source{
The data set comes from the French National Statistics Institute and
the National Geography Institute. It is accessible at
\url{https://www.data.gouv.fr/fr/datasets/contours-iris-insee-ign/},
under the "Licence Ouverte / Open Licence v2.0" licence.
}
\usage{
paris
}
\description{
Geographical data of the French statistical units "IRIS" (\emph{grouped islets for statistical information}) forming the city of Paris.
Five iris units (forming two parcs: that of Boulogne and Vincennes) of dispropotionate size have been removed from the data set.
}
\examples{
\dontrun{
library(graphseg)
## Using "plot"
plot(paris["geometry"])
## Using "ggplot"
library(ggplot)
ggplot(paris) + geom_sf()
}
}
\keyword{datasets}
