% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{solve_spd}
\alias{solve_spd}
\title{Linear solver when the matrix is symmetric positive definite}
\usage{
solve_spd(mat, vect)
}
\arguments{
\item{mat}{the matrix}

\item{vect}{the vector}
}
\value{
\code{mat}\eqn{^{-1}} \code{vec}
}
\description{
Linear solver when the matrix is symmetric positive definite
}
