% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{solver}
\alias{solver}
\title{Compute the weighted ridge penalized estimate}
\usage{
solver(gamma, sigma_sq, K, pen)
}
\arguments{
\item{gamma}{observed signal on a graph}

\item{sigma_sq}{variance of \code{gamma} (functions as weights for \code{gamma})}

\item{K}{weighted laplacian matrix of the graph}

\item{pen}{penalty constant (must be a positive number). Higher values of \code{pen} yields a more regularized fit}
}
\description{
Compute the weighted ridge penalized estimate
}
\seealso{
\code{\link{loglik}}, \code{\link{score}}, \code{\link{hessian}}
}
