% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_old.R
\name{str_detect_2}
\alias{str_detect_2}
\title{Vectorized \code{\link{str_detect}} from package stringr}
\usage{
str_detect_2(string, pattern_vect)
}
\arguments{
\item{string}{string to check for presence of the pattern}

\item{pattern_vect}{pattern to detec}
}
\value{
A vector of boolean of the same length as \code{string}
}
\description{
Vectorized \code{\link{str_detect}} from package stringr
}
\seealso{
\code{\link{str_detect}} from package \code{stringr}
}
\author{
Vivien Goepp \email{vivien.goepp@gmail.com}
}
