% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utrecht_district}
\alias{utrecht_district}
\alias{graph_utrecht_district}
\title{Administrative areas of the Netherlands aroung the city of Utrecht}
\format{
\describe{
  \item{utrecht_district}{An \code{\link[sf]{sf}} object, with lattitude and longitude. With
  7 variables and 650 zones.}
  \item{graph_utrecht_district}{The adjacency graph as an \code{\link{igraph}} object.}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 650 rows and 7 columns.

An object of class \code{igraph} of length 650.
}
\source{
The data set comes from \url{https://geodata.nationaalgeoregister.nl/}.
}
\usage{
utrecht_district

graph_utrecht_district
}
\description{
Geographical data of the Dutch administrative units "District".
}
\keyword{datasets}
