% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphseg-package.R
\docType{package}
\name{graphseg}
\alias{graphseg}
\title{graphseg: A package for piece-wise constant estimation of graph-based signal}
\description{
The main function of this package is \code{agraph}, which performs segmentation
of signals on graphs into a piecewise-constant signal.
(the function \code{graph_aridge} is deprecated and kept here for backward
compatibility).
}
\seealso{
The \code{\link{flsa}} package on CRAN, which performs a similar task
}
